Speed up initializing SandiaDecayDataBase; add sandia.decay.min.xml.
For data that includes coincidences, implemented optimistic check that the "id" for each decay particle cooresponds to the index of the particle in XML file for that transition.  Modified minimize_xml_file.cpp so this would be the case for minimized XML.
Switched to using `atof` instead of `strtof` for parsing floats; however, will still fall back to `strtof` for float values of zero; this speeds up initializing time about 20%.

Added sandia.decay.min.xml, which is the base XML data file, but with tag names minimized, coincidences kept in (but accruacy limited to somethign reasonable).