Rename `SandiaDecay::Nuclide::operator<(rhs)` to `Nuclide::lessThanByDecay(lhs,rhs)`, and add `Nuclide::greaterThanForOrdering(lhs,rhs)`.
The `Nuclide::operator<` ordered nuclides roughly by what decayed into what, or if not related, by which is heavier.
This sorting apparently produced stable results when calling `std::sort` or `std::lower_bound` (or I couldnt find a failing case with lots and lots of differenct inputs and permustations), but it wasnt weakly strickly ordered, in the sense if `A < B`, and `B < C`, then there were some cases where A was not less than C - which is a potential issue.
So switched to  using `Nuclide::greaterThanForOrdering(lhs,rhs)` for ordering Nuclides in vectors, which should be strickly ordered.